requirements.txt
streamlit
pandas
plotly.express
xlrd
openpyxl
streamlit_extras
pillow